[gst][gst-plugin-good]:timeout immediately if pack seq is not continuous
BUM-794
In our scenario, there is no need to route packets when they are transmitted.
In the absence of NACK, if the data packets are found to be discontinuous,
it is clear that these packets have been lost, and there is no need for the
timer to make a timeout judgment.

Signed-off-by: jinjun <jinjun@pinecone.net>